login flow | add relations | work Authurize | jwt

+++ controllers/feed.js
@@ -3,7 +3,7 @@ const fs = require("fs");
 const path = require("path");
 
 const Post = require("../models/post");
-const post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPost = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -17,6 +17,7 @@ exports.getPost = (req, res, next) => {
     })
     .then((posts) => {
       res.status(200).json({
+        id: req.userId,
         posts: posts,
         totalItems: totalPost,
       });
@@ -50,19 +51,28 @@ exports.createPost = (req, res, next) => {
   const image = req.file.path.replace(/\\/g, "/");
   const title = req.body.title;
   const content = req.body.content;
-
+  let creator;
   const post = new Post({
     title: title,
     content: content,
-    creator: { name: "MohamadReza" },
+    creator: req.userId,
     imageUrl: image,
   });
   post
     .save()
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(() => {
       res.status(201).json({
         message: "Product create successfully",
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -104,9 +114,17 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      // Authorization
+      if (req.userId != post.creator.toString()) {
+        const error = new Error("Not Authurize !");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (image !== post.imageUrl) {
         clearImag(post.imageUrl);
       }
+
       post.title = title;
       post.content = content;
       post.imageUrl = image;
@@ -159,10 +177,21 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      // Authorization
+      if (req.userId != post.creator.toString()) {
+        const error = new Error("Not Authurize !");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImag(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
-    .then((result) => {
+    .then(() => User.findById(req.userId))
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(() => {
       res.status(200).json({
         message: "Delete Successfull",
       });
